(Nearly) Finished Product
I say nearly because who knows what I’ll find next wrong with it.
Comments are added at each step of the way. README will be updated
shortly.

Integrator.java is in the src file. This is the only file required to
run the program, since I found a way not to use all the other classes I
wrote. However, they are all still available in /Integrator/src if
people want to check them out. Tree.java was particularly interesting
to write. :)